+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '1.8.9';
+const VERSION = '1.8.10';
 
 /**
  * Display copyright on the admin dashboard.